简化代码
1.优化枚举
2.id创建使用默认配置：雪花id

+++ src/main/java/com/mck/activiti/common/util/TreeUtil.java
@@ -12,7 +12,7 @@ import java.util.List;
  **/
 public class TreeUtil {
 
-    public static List<MenuVo> toTree(List<MenuVo> treeList, Integer pid) {
+    public static List<MenuVo> toTree(List<MenuVo> treeList, Long pid) {
         List<MenuVo> retList = new ArrayList<MenuVo>();
         for (MenuVo parent : treeList) {
             if (pid.equals(parent.getPid())) {
